projects/ad469x_fmc: wire extra GP0 connections
The AD469x FMC board has 3 GP0 connections to the FPGA. Only one was
being used to connect it to the trigger of the SPI Engine offload.

This adds GPIOs connected to the other two pins so that we can
use them for other purposes, e.g. a BUSY or ALERT interrupt.

Signed-off-by: David Lechner <dlechner@baylibre.com>