projects/ad738x_fmc: add DMA gate to SPI offload trigger
Add an AND gate to disable the SPI offload trigger when the DMA is full.

Before this change, when NUM_OF_SDI=1, the SPI offload would pause in the
middle of a SPI message when the DMA got full which confused both the DMA
controller and the ADC chip causing data channels to become misalligned in
memory.

After this change, the trigger is prevented from firing when the DMA is full
so the SPI message doesn't even start, avoiding the problems mentioned above.